App/Backend/requirements.txt
flask
openai
scrapy
flask_cors
langchain
langchain_openai
python-dotenv
assemblyai
sentence_transformers
rake_nltk
matplotlib
scikit-learn
google-cloud-texttospeech